More 64 bit safety

+++ src/include/OpenImageIO/deepdata.h
@@ -155,13 +155,13 @@ public:
     /// same channel layout. Return `true` if ok, `false` if the operation
     /// could not be performed.
     bool copy_deep_sample(int64_t pixel, int sample, const DeepData& src,
-                          int srcpixel, int srcsample);
+                          int64_t srcpixel, int srcsample);
 
     /// Copy an entire deep pixel from `src` to this `DeepData`, completely
     /// eplacing any pixel data for that pixel. They must have the same
     /// channel ayout. Return `true` if ok, `false` if the operation could
     /// not be erformed.
-    bool copy_deep_pixel(int64_t pixel, const DeepData& src, int srcpixel);
+    bool copy_deep_pixel(int64_t pixel, const DeepData& src, int64_t srcpixel);
 
     /// Split all samples of that pixel at the given depth zsplit. Samples
     /// that span z (i.e. z < zsplit < zback) will be split into two samples
